Remove include/bpf_helpers.h and include/bpf_helper_defs.h
Use bpf_helpers.h and bpf_helper_defs.h of libbpf directly.

And 6f8e021c3c8904d19b659898e4b8b83782ef62bb of libbpf had changed
return type of bpf helpers from int to long. So fix redbpf-probes to
convert i64 to i32.

+++ redbpf-probes/src/helpers.rs
@@ -104,7 +104,8 @@ pub unsafe fn bpf_probe_read<T>(src: *const T) -> Result<T, i32> {
         src as *const c_void,
     );
     if ret < 0 {
-        return Err(ret);
+        // TODO return i64
+        return Err(ret as i32);
     }
 
     Ok(v.assume_init())
